Refactored MCP tools
